[WFS provider] Extract reusable cache and background downloader logic
Funded by Planet
