+++ src/providers/wfs/qgswfsutils.cpp
@@ -16,7 +16,6 @@
 #include "qgsapplication.h"
 #include "qgslogger.h"
 #include "qgswfsutils.h"
-#include "qgsgeometry.h"
 #include "qgssettings.h"
 
 // 1 minute
@@ -322,47 +321,3 @@ QString QgsWFSUtils::nameSpacePrefix( const QString &tname )
   return splitList.at( 0 );
 }
 
-
-QString QgsWFSUtils::getMD5( const QgsFeature &f )
-{
-  const QgsAttributes attrs = f.attributes();
-  QCryptographicHash hash( QCryptographicHash::Md5 );
-  for ( int i = 0; i < attrs.size(); i++ )
-  {
-    const QVariant &v = attrs[i];
-    hash.addData( QByteArray( ( const char * )&i, sizeof( i ) ) );
-    if ( v.isNull() )
-    {
-      // nothing to do
-    }
-    else if ( v.type() == QVariant::DateTime )
-    {
-      qint64 val = v.toDateTime().toMSecsSinceEpoch();
-      hash.addData( QByteArray( ( const char * )&val, sizeof( val ) ) );
-    }
-    else if ( v.type() == QVariant::Int )
-    {
-      int val = v.toInt();
-      hash.addData( QByteArray( ( const char * )&val, sizeof( val ) ) );
-    }
-    else if ( v.type() == QVariant::LongLong )
-    {
-      qint64 val = v.toLongLong();
-      hash.addData( QByteArray( ( const char * )&val, sizeof( val ) ) );
-    }
-    else  if ( v.type() == QVariant::String )
-    {
-      hash.addData( v.toByteArray() );
-    }
-  }
-
-  const int attrCount = attrs.size();
-  hash.addData( QByteArray( ( const char * )&attrCount, sizeof( attrCount ) ) );
-  QgsGeometry geometry = f.geometry();
-  if ( !geometry.isNull() )
-  {
-    hash.addData( geometry.asWkb() );
-  }
-
-  return hash.result().toHex();
-}
