+++ src/providers/wfs/qgswfsutils.h
@@ -42,9 +42,6 @@ class QgsWFSUtils
     //! Returns namespace prefix (or an empty string if there is no prefix)
     static QString nameSpacePrefix( const QString &tname );
 
-    //! Returns a unique identifier made from feature content
-    static QString getMD5( const QgsFeature &f );
-
   protected:
     friend class QgsWFSUtilsKeepAlive;
     static QSharedMemory *createAndAttachSHM();
